Add skip to content link
As with best practice for accessibility, this site should have a skip to
content link.

This adds the styles, the link and makes the main content area 
focussable in order for the link to do anything! It's not the most 
perfect implementation but it does work well for the requirements on
this playground site.